requirements.txt
django
django-ninja
python-jose
passlib
python-dotenv
psycopg[binary]
django-cors-headers
